Add annotation for DockerSELinuxMCS bug workaround
This introduces an annotation in the KubeVirt CR to
handle the DockerSELinuxMCS bug: https://github.com/moby/moby/issues/41370.

Using an annotation instead of a dedicated field avoids
locking this workaround into the v1 API, as the bug will
likely be fixed in the future.
Once resolved, we can simply remove the annotation
without impacting the API.

Signed-off-by: bmordeha <bmordeha@redhat.com>